* Change dependencies to testCompile * Move Counter sample to unit test
