+++ sdks/java/io/cdc/build.gradle
@@ -50,11 +50,11 @@ dependencies {
     compile "org.apache.kafka:connect-api:2.5.0"
     compile "org.apache.kafka:connect-json:2.5.0"
 
-    // Debezium depenedencies
+    // Debezium dependencies
     compile group: 'io.debezium', name: 'debezium-core', version: '1.3.1.Final'
-    compile group: 'io.debezium', name: 'debezium-connector-mysql', version: '1.3.1.Final'
-    compile group: 'io.debezium', name: 'debezium-connector-sqlserver', version: '1.3.1.Final'
-    compile group: 'io.debezium', name: 'debezium-connector-postgres', version: '1.3.1.Final'
+    testCompile group: 'io.debezium', name: 'debezium-connector-mysql', version: '1.3.1.Final'
+    testCompile group: 'io.debezium', name: 'debezium-connector-sqlserver', version: '1.3.1.Final'
+    testCompile group: 'io.debezium', name: 'debezium-connector-postgres', version: '1.3.1.Final'
 }
 
 test {
